Session Not Found
If a user browses away from the page and then hits the browser back button, the browser reuses the session cookie, but the connection for that session was closed when they browsed away.

If the session is not found when a WebSocket connection is made, this fix will force the browser to reload the page. The reload will start a new session and give the user a new cookie.

- Delete session on disconnect
- Recover from panics that happen on WebSocket requests
- More error checking
- More logging
- Add ability to send errors to the browser via WebSocket